+++ pyproject.toml
@@ -93,7 +93,7 @@ mssql = [
     "pyodbc>=5.0.0",
 ]
 document = [
-    "pypdfium2>=4.30.1",      # PDF documents
+    "pypdf>=5.2.1",           # PDF documents
     "python-docx>=1.2.0",     # Word documents
     "python-pptx>=0.6.23",    # PowerPoint presentations
     "PyYAML>=6.0.1",          # YAML files
