+++ src/toolfront/models/actions/read.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel, Field
+
+from toolfront.models.connections.library import LibraryConnection
+
+
+class Read(BaseModel):
+    """Storage document."""
+
+    connection: LibraryConnection = Field(..., description="Library connection.")
