+++ src/toolfront/models/actions/request.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from pydantic import BaseModel, Field
 
-from toolfront.models.connection import APIConnection
+from toolfront.models.connections.api import APIConnection
 from toolfront.types import HTTPMethod
