+++ src/toolfront/models/databases/mysql.py
@@ -2,9 +2,9 @@ from typing import Any
 
 import pandas as pd
 
+from toolfront.cache import cache
 from toolfront.config import CACHE_TTL
 from toolfront.models.database import Database, SQLAlchemyMixin
-from toolfront.storage import cache
 
 
 class MySQL(SQLAlchemyMixin, Database):
